requirements.txt
Flask
gunicorn
pyTelegramBotAPI
gspread
oauth2client
telethon
google-generativeai